Adding the testing for User Service with its exceptions

+++ service/src/main/java/hu/uni/miskolc/iit/service/UserServiceImpl.java
@@ -1,14 +1,16 @@
 package hu.uni.miskolc.iit.service;
-import hu.uni.miskolc.iit.exceptions.ListNotFoundException;
-import hu.uni.miskolc.iit.exceptions.UserErrorException;
-import hu.uni.miskolc.iit.exceptions.EntryNotFoundException;
+
+import hu.uni.miskolc.iit.dao.DAOimpl.UserDAOImpl;
 import hu.uni.miskolc.iit.dao.UserDAO;
+import hu.uni.miskolc.iit.exceptions.UserErrorException;
 import hu.uni.miskolc.iit.model.User;
 
 
 public class UserServiceImpl implements UserService {
 
-    private UserDAO userdao;
+    private UserDAO userdao = new UserDAOImpl();
+
+    public UserServiceImpl(){}
 
     public UserServiceImpl(UserDAO userdao) {
         this.userdao = userdao;
@@ -16,28 +18,26 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User userRegistration(User user) throws UserErrorException {
-        try {
-            return userdao.createUser(user);
-        } catch (UserErrorException e) {
+        if(user == null) {
             throw new UserErrorException();
         }
+        return userdao.createUser(user);
+
     }
 
     @Override
-    public User userModification(User user) throws EntryNotFoundException {
-        try {
+    public User userModification(User user) throws UserErrorException {
+        if(user == null) {
+            throw new UserErrorException();
+        }
             return userdao.updateUser(user);
-        } catch (EntryNotFoundException e) {
-            throw new EntryNotFoundException();
-        }
     }
 
     @Override
-    public void userDelete(User user) throws EntryNotFoundException {
-        try {
-            userdao.deleteUser(user);
-        } catch (EntryNotFoundException e) {
-            throw new EntryNotFoundException();
+    public boolean userDelete(User user) throws UserErrorException {
+        if(user == null) {
+            throw new UserErrorException();
         }
+        return  userdao.deleteUser(user);
     }
 }
